Remove v1 protocol file. Move all clearly protocol related functions into V0 folder.
