+++ protocol-models/src/main/java/io/airbyte/protocol/models/v0/Field.java
@@ -2,7 +2,9 @@
  * Copyright (c) 2020-2025 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.protocol.models;
+package io.airbyte.protocol.models.v0;
+
+import io.airbyte.protocol.models.JsonSchemaType;
 
 import java.util.List;
